+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Add support `filter` in `EveOnline::ESI::UniverseStructures`
 * Add support `max_war_id` in `EveOnline::ESI::Wars`
 * Add support `from_event` in `EveOnline::ESI::CharacterCalendar`
+* Add `EveOnline::ESI::CorporationAllianceHistory`
+* Add `EveOnline::ESI::Models::CorporationAllianceHistory`
 
 ## v0.27.0